#2225 :recycle: combine ergebnis + ergebnisermittlung mappers
